Moved to dictionary
*added role check
*moved components to dictionary
