+++ src/app/modules/shared/components/dictionaries/manufacturer/delete-manufacturer/delete-manufacturer.component.html
