+++ src/app/modules/shared/components/dictionaries/manufacturer/manufacturer.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ManufacturerService } from 'src/app/modules/shared/services/manufacturer.service';
 import { Manufacturer } from 'src/app/modules/shared/models/manufacturer';
-import { EntitiesDataSource } from '../../data-sources/entities-data-sourse';
-import { MatFspTableComponent } from '../tables/mat-fsp-table/mat-fsp-table.component';
+import { EntitiesDataSource } from '../../../data-sources/entities-data-sourse';
+import { MatFspTableComponent } from '../../tables/mat-fsp-table/mat-fsp-table.component';
+import { AuthenticationService } from 'src/app/modules/core/services/authentication.service';
 
 @Component({
   selector: 'app-manufacturer',
@@ -10,6 +11,7 @@ import { MatFspTableComponent } from '../tables/mat-fsp-table/mat-fsp-table.comp
   styleUrls: ['./manufacturer.component.scss']
 })
 export class ManufacturerComponent implements OnInit {
+  isAdmin: boolean;
   columnDefinitions: string[] = [
     'name',
   ];
@@ -21,11 +23,12 @@ export class ManufacturerComponent implements OnInit {
 
   dataSource: EntitiesDataSource<Manufacturer>;
 
-  constructor(private manufacturerService: ManufacturerService) {
+  constructor(private manufacturerService: ManufacturerService, private authenticationService: AuthenticationService) {
   }
 
   ngOnInit() {
     this.dataSource = new EntitiesDataSource<Manufacturer>(this.manufacturerService);
+    this.isAdmin = this.authenticationService.getRole() === 'ADMIN' ? true : false;
   }
 
   addManufacturer(manufacturer: Manufacturer) {
