+++ src/app/modules/shared/components/dictionaries/unit/unit.component.html
@@ -1,0 +1,17 @@
+<div class="table-background bg-white border">
+  <app-create-unit *ngIf="isAdmin" (createUnit)="addUnit($event)"></app-create-unit>
+
+  <ng-template #actionsTemplate let-data="templateData">
+
+    <app-delete-unit *ngIf="isAdmin" [unit]="data" (deleteUnit)="deleteUnit($event)"></app-delete-unit>
+
+    <app-edit-unit *ngIf="isAdmin" [unit]="data" (updateUnit)="updateUnit($event)"></app-edit-unit>
+
+  </ng-template>
+
+  <div>
+    <mat-fsp-table #table [columnDefinitions]="columnDefinitions" [columnNames]="columnNames" [dataSource]="dataSource"
+      [actionContentTemplate]="actionsTemplate">
+    </mat-fsp-table>
+  </div>
+</div>
