--- src/app/modules/shared/components/unit/unit.component.html
@@ -1,25 +1,0 @@
-<app-create-unit (createUnit)="addUnit($event)"></app-create-unit>
-
-<ng-template #actionsTemplate let-data="templateData">
-
-<app-delete-unit
-    [unit]="data"
-    (deleteUnit)="deleteUnit($event)" 
-></app-delete-unit>
-
-<app-edit-unit
-    [unit]="data"
-    (updateUnit)="updateUnit($event)"
-></app-edit-unit>
-
-</ng-template>
-
-<div>
-    <mat-fsp-table #table
-        [columnDefinitions]="columnDefinitions"
-        [columnNames]="columnNames"
-        [dataSource]="dataSource"
-        [actionContentTemplate]="actionsTemplate">   
-    </mat-fsp-table>
-</div>
-
