+++ src/assets/translate/shared/manufacturer/uk.json
@@ -2,7 +2,7 @@
     "Manufacturer": {
         "name": "Назва",
         "Create": {
-            "Title": "Створити виробника"
+            "Title": "Додати виробника"
         },
         "Delete": {
             "Title": "Ви дійсно хочeте видалити виробника?",
