+++ src/assets/translate/shared/unit/uk.json
@@ -3,7 +3,7 @@
         "name": "Назва",
         "shortName": "Коротка назва",
         "Create": {
-            "Title": "Створити одиницю вимірювання"
+            "Title": "Додати одиницю вимірювання"
         },
         "Delete": {
             "Title": "Ви дійсно хочeте видалити одиницю вимірювання?",
